refactor: move numeric input formatting utility functions
